test: address scheduler review nits
