+++ operator/internal/clustertopology/clustertopology_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/ai-dynamo/grove/operator/internal/scheduler"
 	"github.com/ai-dynamo/grove/operator/internal/scheduler/kai"
 	testutils "github.com/ai-dynamo/grove/operator/test/utils"
-	"github.com/ai-dynamo/grove/operator/test/utils/schedulertest"
+	schedulertest "github.com/ai-dynamo/grove/operator/test/utils/scheduler"
 
 	"github.com/go-logr/logr"
 	kaitopologyv1alpha1 "github.com/kai-scheduler/KAI-scheduler/pkg/apis/kai/v1alpha1"
@@ -41,14 +41,6 @@ import (
 
 const topologyName = "test-topology"
 
-func newKAIBackends(t *testing.T, cl client.Client) map[string]scheduler.TopologyAwareBackend {
-	t.Helper()
-	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameKai}
-	b := kai.New(cl, cl.Scheme(), nil, profile)
-	require.NoError(t, b.Init(cl))
-	return map[string]scheduler.TopologyAwareBackend{b.Name(): b.(scheduler.TopologyAwareBackend)}
-}
-
 func TestSynchronizeTopologyListsAndSyncs(t *testing.T) {
 	ctx := context.Background()
 	topologyLevels := []grovecorev1alpha1.TopologyLevel{
@@ -60,7 +52,7 @@ func TestSynchronizeTopologyListsAndSyncs(t *testing.T) {
 	cl := schedulertest.NewKAIClient(t, ct)
 	logger := logr.Discard()
 
-	err := SynchronizeTopology(ctx, cl, logger, newKAIBackends(t, cl))
+	err := SynchronizeTopology(ctx, cl, logger, newKAIBackendMap(t, cl))
 	require.NoError(t, err)
 
 	// Verify KAI Topology was created
@@ -85,7 +77,7 @@ func TestSynchronizeTopologyMultipleCTs(t *testing.T) {
 	cl := schedulertest.NewKAIClient(t, ct1, ct2)
 	logger := logr.Discard()
 
-	err := SynchronizeTopology(ctx, cl, logger, newKAIBackends(t, cl))
+	err := SynchronizeTopology(ctx, cl, logger, newKAIBackendMap(t, cl))
 	require.NoError(t, err)
 
 	// Verify both KAI Topologies were created
@@ -105,7 +97,7 @@ func TestSynchronizeTopologyNoCTs(t *testing.T) {
 	cl := schedulertest.NewKAIClient(t)
 	logger := logr.Discard()
 
-	err := SynchronizeTopology(ctx, cl, logger, newKAIBackends(t, cl))
+	err := SynchronizeTopology(ctx, cl, logger, newKAIBackendMap(t, cl))
 	require.NoError(t, err)
 }
 
@@ -139,7 +131,7 @@ func TestSynchronizeTopologySkipsExternallyManaged(t *testing.T) {
 	// KAI backend is listed in schedulerTopologyReferences — SyncTopology should NOT be called.
 	// If it were called, it would try to create a KAI Topology and succeed, so we verify
 	// that no KAI Topology was created.
-	err := SynchronizeTopology(ctx, cl, logger, newKAIBackends(t, cl))
+	err := SynchronizeTopology(ctx, cl, logger, newKAIBackendMap(t, cl))
 	require.NoError(t, err)
 
 	kaiTopology := &kaitopologyv1alpha1.Topology{}
@@ -157,7 +149,7 @@ func TestSynchronizeTopologyListError(t *testing.T) {
 		Build()
 	logger := logr.Discard()
 
-	err := SynchronizeTopology(ctx, cl, logger, newKAIBackends(t, cl))
+	err := SynchronizeTopology(ctx, cl, logger, newKAIBackendMap(t, cl))
 	assert.Error(t, err)
 }
 
@@ -292,8 +284,26 @@ func TestGetClusterTopologyLevels(t *testing.T) {
 	}
 }
 
-// Helper functions for creating test resources
-// --------------------------------------------------
+func TestBuildSchedulerReferenceMap(t *testing.T) {
+	refs := []grovecorev1alpha1.SchedulerTopologyBinding{
+		{SchedulerName: "kai-scheduler", TopologyReference: "kai-topo"},
+		{SchedulerName: "other-scheduler", TopologyReference: "other-topo"},
+	}
+	m := BuildSchedulerReferenceMap(refs)
+	assert.NotNil(t, m["kai-scheduler"])
+	assert.Equal(t, "kai-topo", m["kai-scheduler"].TopologyReference)
+	assert.Nil(t, m["nonexistent"])
+	assert.Empty(t, BuildSchedulerReferenceMap(nil))
+}
+
+func newKAIBackendMap(t *testing.T, cl client.Client) map[string]scheduler.TopologyAwareBackend {
+	t.Helper()
+	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameKai}
+	b := kai.New(cl, cl.Scheme(), nil, profile)
+	require.NoError(t, b.Init(cl))
+	return map[string]scheduler.TopologyAwareBackend{b.Name(): b.(scheduler.TopologyAwareBackend)}
+}
+
 // createTestClusterTopology creates a ClusterTopologyBinding with the given name and topology levels.
 func createTestClusterTopology(name string, levels []grovecorev1alpha1.TopologyLevel) *grovecorev1alpha1.ClusterTopologyBinding {
 	return &grovecorev1alpha1.ClusterTopologyBinding{
@@ -305,16 +315,4 @@ func createTestClusterTopology(name string, levels []grovecorev1alpha1.TopologyL
 			Levels: levels,
 		},
 	}
-}
-
-func TestBuildSchedulerReferenceMap(t *testing.T) {
-	refs := []grovecorev1alpha1.SchedulerTopologyBinding{
-		{SchedulerName: "kai-scheduler", TopologyReference: "kai-topo"},
-		{SchedulerName: "other-scheduler", TopologyReference: "other-topo"},
-	}
-	m := BuildSchedulerReferenceMap(refs)
-	assert.NotNil(t, m["kai-scheduler"])
-	assert.Equal(t, "kai-topo", m["kai-scheduler"].TopologyReference)
-	assert.Nil(t, m["nonexistent"])
-	assert.Empty(t, BuildSchedulerReferenceMap(nil))
 }
