+++ operator/internal/scheduler/registry/registry_test.go
@@ -22,24 +22,12 @@ import (
 	configv1alpha1 "github.com/ai-dynamo/grove/operator/api/config/v1alpha1"
 	"github.com/ai-dynamo/grove/operator/internal/scheduler"
 	testutils "github.com/ai-dynamo/grove/operator/test/utils"
+	schedulertest "github.com/ai-dynamo/grove/operator/test/utils/scheduler"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-func newVolcanoDirectClient(t *testing.T, objects ...client.Object) client.Client {
-	t.Helper()
-	scheme := runtime.NewScheme()
-	require.NoError(t, apiextensionsv1.AddToScheme(scheme))
-	return testutils.NewTestClientBuilder().
-		WithScheme(scheme).
-		WithObjects(objects...).
-		Build()
-}
 
 // TestNewRegistry tests New with different scheduler profiles.
 func TestNewRegistry(t *testing.T) {
@@ -106,7 +94,7 @@ func TestNewRegistry(t *testing.T) {
 	}
 
 	t.Run("multiple profiles with default set to kai", func(t *testing.T) {
-		cl := testutils.CreateDefaultFakeClient([]client.Object{testutils.NewVolcanoPodGroupCRD(true)})
+		cl := schedulertest.NewVolcanoClient(t, testutils.NewVolcanoPodGroupCRD(true))
 		recorder := record.NewFakeRecorder(10)
 		cfg := configv1alpha1.SchedulerConfiguration{
 			Profiles: []configv1alpha1.SchedulerProfile{
@@ -128,11 +116,7 @@ func TestNewRegistry(t *testing.T) {
 	})
 
 	t.Run("volcano scheduler initialization", func(t *testing.T) {
-		scheme := runtime.NewScheme()
-		cl := testutils.NewTestClientBuilder().
-			WithScheme(scheme).
-			Build()
-		directClient := newVolcanoDirectClient(t, testutils.NewVolcanoPodGroupCRD(true))
+		cl := schedulertest.NewVolcanoClient(t, testutils.NewVolcanoPodGroupCRD(true))
 
 		recorder := record.NewFakeRecorder(10)
 		cfg := configv1alpha1.SchedulerConfiguration{
@@ -141,7 +125,7 @@ func TestNewRegistry(t *testing.T) {
 			},
 			DefaultProfileName: string(configv1alpha1.SchedulerNameVolcano),
 		}
-		reg, err := New(cl, directClient, scheme, recorder, cfg)
+		reg, err := New(cl, cl, cl.Scheme(), recorder, cfg)
 		require.NoError(t, err)
 		require.NotNil(t, reg.GetDefault())
 		assert.Equal(t, string(configv1alpha1.SchedulerNameVolcano), reg.GetDefault().Name())
