+++ operator/internal/scheduler/volcano/backend_test.go
@@ -24,38 +24,20 @@ import (
 	configv1alpha1 "github.com/ai-dynamo/grove/operator/api/config/v1alpha1"
 	grovecorev1alpha1 "github.com/ai-dynamo/grove/operator/api/core/v1alpha1"
 	testutils "github.com/ai-dynamo/grove/operator/test/utils"
+	schedulertest "github.com/ai-dynamo/grove/operator/test/utils/scheduler"
 
 	groveschedulerv1alpha1 "github.com/ai-dynamo/grove/scheduler/api/core/v1alpha1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	volcanov1beta1 "volcano.sh/apis/pkg/apis/scheduling/v1beta1"
 )
 
-func newVolcanoScheme(t *testing.T) *runtime.Scheme {
-	t.Helper()
-	scheme := runtime.NewScheme()
-	require.NoError(t, groveschedulerv1alpha1.AddToScheme(scheme))
-	require.NoError(t, volcanov1beta1.AddToScheme(scheme))
-	require.NoError(t, apiextensionsv1.AddToScheme(scheme))
-	return scheme
-}
-
-func newVolcanoClient(t *testing.T, objects ...client.Object) client.Client {
-	t.Helper()
-	return testutils.NewTestClientBuilder().
-		WithScheme(newVolcanoScheme(t)).
-		WithObjects(objects...).
-		Build()
-}
-
 func TestBackend_PreparePod(t *testing.T) {
-	cl := newVolcanoClient(t)
+	cl := schedulertest.NewVolcanoClient(t)
 	recorder := record.NewFakeRecorder(10)
 	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 	b := New(cl, cl.Scheme(), recorder, profile)
@@ -71,7 +53,7 @@ func TestBackend_PreparePod(t *testing.T) {
 }
 
 func TestBackend_PreparePodFailsWhenPodGangLabelMissing(t *testing.T) {
-	cl := newVolcanoClient(t)
+	cl := schedulertest.NewVolcanoClient(t)
 	recorder := record.NewFakeRecorder(10)
 	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 	b := New(cl, cl.Scheme(), recorder, profile)
@@ -104,7 +86,7 @@ func TestBackend_Init(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			cl := newVolcanoClient(t, tt.objects...)
+			cl := schedulertest.NewVolcanoClient(t, tt.objects...)
 			recorder := record.NewFakeRecorder(10)
 			profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 			b := New(cl, cl.Scheme(), recorder, profile)
@@ -141,7 +123,7 @@ func TestBackend_SyncPodGang(t *testing.T) {
 			},
 		},
 	}
-	cl := newVolcanoClient(t, podGang)
+	cl := schedulertest.NewVolcanoClient(t, podGang)
 	recorder := record.NewFakeRecorder(10)
 	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 	b := New(cl, cl.Scheme(), recorder, profile)
@@ -188,7 +170,7 @@ func TestBackend_SyncPodGangPreservesSchedulingConstraintsAfterRelease(t *testin
 			},
 		},
 	}
-	cl := newVolcanoClient(t, podGang)
+	cl := schedulertest.NewVolcanoClient(t, podGang)
 	recorder := record.NewFakeRecorder(10)
 	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 	b := New(cl, cl.Scheme(), recorder, profile)
@@ -232,7 +214,7 @@ func TestBackend_SyncPodGangDefaultQueue(t *testing.T) {
 			},
 		},
 	}
-	cl := newVolcanoClient(t, podGang)
+	cl := schedulertest.NewVolcanoClient(t, podGang)
 	recorder := record.NewFakeRecorder(10)
 	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 	b := New(cl, cl.Scheme(), recorder, profile)
@@ -247,7 +229,7 @@ func TestBackend_SyncPodGangDefaultQueue(t *testing.T) {
 }
 
 func TestBackend_ValidatePodCliqueSet(t *testing.T) {
-	cl := newVolcanoClient(t)
+	cl := schedulertest.NewVolcanoClient(t)
 	recorder := record.NewFakeRecorder(10)
 	profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 	b := New(cl, cl.Scheme(), recorder, profile)
@@ -393,7 +375,7 @@ func TestBackend_ValidatePodCliqueSetQueues(t *testing.T) {
 		t.Run(tc.description, func(t *testing.T) {
 			pcs := makePCS()
 			tc.mutate(pcs)
-			cl := newVolcanoClient(t, tc.existingObjs...)
+			cl := schedulertest.NewVolcanoClient(t, tc.existingObjs...)
 			recorder := record.NewFakeRecorder(10)
 			profile := configv1alpha1.SchedulerProfile{Name: configv1alpha1.SchedulerNameVolcano}
 			b := New(cl, cl.Scheme(), recorder, profile)
