+++ operator/test/utils/scheduler/kai.go
@@ -1,0 +1,47 @@
+// /*
+// Copyright 2026 The Grove Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// */
+
+package scheduler
+
+import (
+	"testing"
+
+	grovecorev1alpha1 "github.com/ai-dynamo/grove/operator/api/core/v1alpha1"
+	testutils "github.com/ai-dynamo/grove/operator/test/utils"
+
+	kaitopologyv1alpha1 "github.com/kai-scheduler/KAI-scheduler/pkg/apis/kai/v1alpha1"
+	"github.com/stretchr/testify/require"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// NewKAIScheme returns an isolated scheme with the Grove and KAI types used by KAI scheduler tests.
+func NewKAIScheme(t *testing.T) *runtime.Scheme {
+	t.Helper()
+	scheme := runtime.NewScheme()
+	require.NoError(t, grovecorev1alpha1.AddToScheme(scheme))
+	require.NoError(t, kaitopologyv1alpha1.AddToScheme(scheme))
+	return scheme
+}
+
+// NewKAIClient returns a fake client using NewKAIScheme.
+func NewKAIClient(t *testing.T, objects ...client.Object) client.Client {
+	t.Helper()
+	return testutils.NewTestClientBuilder().
+		WithScheme(NewKAIScheme(t)).
+		WithObjects(objects...).
+		Build()
+}
